add service to save in db transactional

+++ src/main/java/com/kpei/mkbd/datatransform/dto/Vd55Dto.java
@@ -2,6 +2,8 @@ package com.kpei.mkbd.datatransform.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
@@ -9,11 +11,11 @@ import lombok.*;
 @NoArgsConstructor
 public class Vd55Dto extends Vd5Base {
     private String kodeEfek;
-    private Double nilaiEfek;
+    private BigDecimal nilaiEfek;
     private String kodeEfekLn;
-    private Double nilaiEfekLn;
-    private Double nilaiEfekTutupLn;
-    private Double nilaiHaircutTutupLn;
-    private Double nilaiHaircutLn;
-    private Double jmlPengembalianHaircut;
+    private BigDecimal nilaiEfekLn;
+    private BigDecimal nilaiEfekTutupLn;
+    private BigDecimal nilaiHaircutTutupLn;
+    private BigDecimal nilaiHaircutLn;
+    private BigDecimal jmlPengembalianHaircut;
 }
